新增云更新Onnx资源功能，优化识别yolo棋盘方式

+++ ChessAI/QClientSocket.cpp
@@ -62,10 +62,10 @@ void QClientSocket::SendCommand(int cmd,unsigned char* data,int size)
 int QClientSocket::DealCommand()
 {
 	char* buffer = m_buffer.data();
-	static size_t index = 0;
+	static int index = 0;
 	while (true)
 	{
-		size_t len = recv(m_sock, buffer + index, BUFFER_SIZE - index, 0);
+		int len = recv(m_sock, buffer + index, BUFFER_SIZE - index, 0);
 		if (len <= 0 && index <=0 )
 		{
 			return -1;
